pyproject.toml
[project]
name = "smartpdf-rag"
version = "0.1.0"
description = "Chatbot RAG utilisant Google Gemini et FAISS pour interroger des PDF"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit",
    "langchain",
    "langchain-google-genai",
    "langchain-community",
    "langchain-text-splitters",
    "faiss-cpu",
    "python-dotenv",
    "pypdf",
    "google-generativeai",
    "langchain-openai>=0.3.35",
]

[dependency-groups]
dev = [
    "black",
    "isort",
]
